add tests to verify game tick behavior
My general thought is to have a GameSession that has GameRounds that
have Verses that have VerseWords and at the top level, a tick! will be
forwarded down the chain to cause a state change whereever the game is
at that point.  Tricky part is avoiding some kind of race condition
where the game never ends.

anyway some rough tests to set things up for ticking away