add restful hasura endpoint for creating a user
Hasura lets you RESTize queries and mutations, this is my
first attempt at it.

Basically $hasura/createUser when POSTED to with a name parameter
should create a user with that name and a uuid.  and return
both name and uuid.

don't really like that the returning payload is called "insert_user_one"
but I don't see an easy way around that.