add game_session_scores table
This table just has the user_uuid and game_session_uuid and a score
(integer)

Yes, this is super primitive.  Any correct answer should simply
insert a score.  Easy to spoof, doesn't track which answer was for
which score.  To get a total score you just add up all the scores
for a user for a session

That's fine though.  We just want something working!