Move hasura metadata out of rails api folder
Hasura is an independent service and the only reason to put it in the
rails folder is to help the rails container apply metadata once it boots
up.

There are other ways to this that are more intuitive (using docker
compose) so I moved the hasura folder into the root of the project
alongisde the other two services (rails / react)

also added optional env overrides for the dockerfile for local dev if
needed.